pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "initrunner"
version = "0.2.0b1"
description = "A lightweight AI agent runner"
readme = "README.md"
license = "MIT"
authors = [{ name = "vladimir kesler", email = "contact@initrunner.ai" }]
keywords = ["ai", "agent", "runner", "llm", "pydantic-ai", "mcp", "rag"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai-slim[openai,fastmcp]>=1.56.0",
    "pydantic>=2.11",
    "typer>=0.21.1",
    "pyyaml>=6.0.3",
    "rich>=14.3.2",
    "sqlite-vec>=0.1.6",
    "watchfiles>=1.1.1",
    "croniter>=6.0.0",
    "beautifulsoup4>=4.13.2",
    "markdownify>=1.2.2",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
ingest = ["pymupdf4llm>=0.2.9", "python-docx>=1.2.0", "openpyxl>=3.1.5"]
local-embeddings = ["fastembed>=0.7.4"]
safety = ["better-profanity>=0.7.0"]
anthropic = ["pydantic-ai-slim[anthropic]>=1.56.0"]
google = ["pydantic-ai-slim[google]>=1.56.0"]
groq = ["pydantic-ai-slim[groq]>=1.56.0"]
mistral = ["pydantic-ai-slim[mistral]>=1.56.0"]
all-models = ["pydantic-ai-slim[anthropic,google,groq,mistral,cohere,bedrock,xai]>=1.56.0"]
tui = ["textual>=7.5.0"]
dashboard = ["fastapi>=0.115.0", "uvicorn[standard]>=0.34.0", "jinja2>=3.1"]

[project.scripts]
initrunner = "initrunner.cli.main:app_entry"

[project.urls]
Repository = "https://github.com/vladkesler/initrunner"
Changelog = "https://github.com/vladkesler/initrunner/blob/main/CHANGELOG.md"
Issues = "https://github.com/vladkesler/initrunner/issues"

[dependency-groups]
dev = [
    "pre-commit>=4.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-timeout>=2.3.1",
    "ruff>=0.15.0",
    "ty>=0.0.15",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "textual>=7.5.0",
    "jinja2>=3.1",
]

[tool.hatch.build.targets.wheel]
packages = ["initrunner"]

[tool.pytest.ini_options]
timeout = 60

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["."]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "RUF",  # ruff-specific
]

[tool.ruff.lint.per-file-ignores]
"initrunner/_yaml.py" = ["UP047"]  # TypeVar needed for Python 3.11 compat (PEP 695 is 3.12+)
"initrunner/tui/**" = ["RUF012"]  # Textual uses mutable class attrs (BINDINGS, CSS_PATH)

[tool.ruff.lint.isort]
known-first-party = ["initrunner"]

[tool.ruff.format]
quote-style = "double"

[tool.ty.environment]
python-version = "3.13"
